Feature: refatorando o código
