+++ src/Metodos.java
@@ -1,9 +1,109 @@
+import classes.*;
+
 import java.util.Random;
 import java.util.Scanner;
 
 /* Funções Auxiliares */
-
 public class Metodos {
+    // Variavel para verificar se o personagem esta ou não defendendo
+    static boolean defendendo = false;
+    
+    // Menu
+    public static void menu() {
+        Scanner scanner = new Scanner(System.in);
+        boolean sair = false;
+
+        // Boas vindas
+        clearConsole(); // Limpa o console
+        System.out.printf("Bem-vindo ao From the Ashes\n");
+        System.out.printf("\nPressione Enter para iniciar o jogo...");
+        scanner.nextLine(); // Avança para a próxima linha
+        
+        // Menu Principal
+        while (!sair) {
+            clearConsole(); // Limpa o console
+            System.out.printf("Escolha uma das opções abaixo: \n\n1. Introdução\n2. Jogar\n3. Créditos\n4. Sair\n");
+            System.out.printf("\n>"); // Indica onde o usuario irá digitar
+            int opcao = scanner.nextInt();
+            scanner.nextLine(); // Avança para a próxima linha
+            
+            switch (opcao) {
+                case 1:
+                    clearConsole(); // Limpa o console
+                    System.out.printf("-- INTRODUÇÃO --\n");
+                    System.out.printf("\nEste jogo se passa em um mundo governado por deuses, onde humanos e criaturas das trevas lutam pela sobrevivência. \nUm grande demônio, conhecido como Rei Demônio, lidera um exército de monstros e demônios das profundezas do submundo, ameaçando o mundo e marcando possivelmente o seu fim. \nPara salvar a humanidade você precisará lutar contra monstros poderosos e se aprofundar cada vez mais neste universo.\n");
+                    System.out.printf("\nDurante a sua aventura, você irá se deparar com diversos monstros que precisam ser derrotados para que você possa avançar na história. \nPara enfrentá-los, você terá a opção de atacar, defender ou até mesmo fugir. \nSuas escolhas durante as batalhas serão cruciais para o resultado final e para o seu progresso na jornada. \n\nPortanto, escolha sabiamente e use suas habilidades e recursos com cautela para vencer essas batalhas épicas. \nBoa sorte!\n\n");
+                    delay(1500); // Adiciona atraso 
+                    System.out.println(" ");
+                    System.out.print("Pressione Enter para continuar...");
+                    scanner.nextLine(); // Avança para a próxima linha
+                    System.out.println(" ");
+                    break;
+                case 2:
+                    clearConsole(); // Limpa o console
+                    jogar(); // Inicia o Jogo
+                    break;
+                case 3:
+                    clearConsole(); // Limpa o console
+                    System.out.printf("--CRÉDITOS--");
+                    delay(1000); // Adiciona atraso 
+                    System.out.println(" ");
+                    System.out.println("Desenvolvido por: Ecthor Silva");
+                    delay(1500); // Adiciona atraso 
+                    System.out.println(" ");
+                    System.out.printf("Pressione Enter para continuar...");
+                    scanner.nextLine(); // Avança para a próxima linha
+                    System.out.println(" ");
+                    break;
+                case 4:
+                    sair = true;
+                    clearConsole(); // Limpa o console
+                    System.out.println("Saindo do jogo...\n");
+                    delay(1500); // Adiciona atraso 
+                    break;
+                default:
+                    clearConsole(); // Limpa o console
+                    System.out.print("Opção inválida, pressione Enter para tentar novamente!");
+                    scanner.nextLine(); // Avança para a próxima linha
+                    System.out.println(" ");
+            }
+        }
+        scanner.close();
+    }
+    // Iniciando o jogo
+    public static void jogar(){
+        Scanner scanner = new Scanner(System.in);
+
+        clearConsole(); // Limpa o console
+        System.out.print("Carregando...");
+        delay(1000); // Adiciona atraso 
+        clearConsole(); // Limpa o console
+        delay(1000); // Adiciona atraso 
+        System.out.printf("Era uma vez um mundo governado por deuses poderosos, onde humanos e criaturas das trevas lutavam pela sobrevivência. \nA humanidade vivia em cidades fortificadas, protegidas por muros altos e magia antiga. No entanto, mesmo com todos esses recursos, uma ameaça se aproximava do horizonte.\n");
+        delay(1000); // Adiciona atraso
+        System.out.println(" ");
+        System.out.printf("Rumores surgiram de que um grande demônio estava se aproximando, liderando um exército de monstros e demônios das profundezas do submundo. \nO demônio era conhecido como o Rei Demônio e muitos acreditavam que sua chegada marcaria o fim do mundo.\n");
+        delay(1000); // Adiciona atraso
+        
+        // Criar personagem
+        Personagem personagem = criarPersonagem(scanner);
+
+        delay(1000);
+        System.out.printf("\n-- ATO I --\n\n");
+        delay(1000);
+        
+        /* Ato I */ 
+        Atos.visitarTaverna(personagem, scanner);
+
+        delay(1000);
+        System.out.printf("\n-- Fim do ATO I --\n");
+        delay(1000);
+        
+        System.out.printf("\nPressione Enter para continuar...");
+        
+        scanner.nextLine(); // Avança para a próxima linha
+        clearConsole(); // Limpa o console
+    }
     // Seleção da Classe do personagem
     public static Personagem criarPersonagem(Scanner scanner) {
         System.out.printf("\nAntes de iniciar você precisará selecionar cuidadosamente sua classe. \nSeu treinamento e habilidades em batalha serão essenciais para superar os desafios que encontrará em sua jornada.\n\n");
@@ -27,7 +127,7 @@ public class Metodos {
                 personagem = new Arqueiro("Arqueiro", "arco", 200, 50, 39, 35, 0); // nome, arma, vida, forca, defesa, destreza e estusFlask
                 break;
             case 4: 
-                personagem = new Clerigo("Clerigo", "castiçal de clérigo", 201, 27, 37, 5, 0); // nome, arma, vida, forca, defesa, destreza e estusFlask
+                personagem = new Clerigo("Clérigo", "castiçal de clérigo", 201, 27, 37, 5, 0); // nome, arma, vida, forca, defesa, destreza e estusFlask
                 break;
             default:
                 System.out.println("Opção inválida, saindo...");
@@ -35,7 +135,7 @@ public class Metodos {
         }
         clearConsole(); // Limpa o console
         System.out.printf("Você escolheu a classe %s.\n", personagem.getNome());
-        if(personagem.getNome().equals("Clerigo")){
+        if(personagem.getNome().equals("Clérigo")){
             System.out.printf("\nCalma, você realmente escolheu essa classe?\nDe todas as outras vc escolheu essa? Boa sorte! kkkkkkkk\n");
         }
         delay(1500); // Adiciona atraso 
@@ -107,7 +207,7 @@ public class Metodos {
                     clearConsole(); // Limpa o console
                     // DEFESA
                     System.out.printf("Você está defendendo do ataque do %s e receberá apenas metade do dano!\n", monstro.getNome());
-                    Game.defendendo = true; // Atribui o valor verdadeiro a variavel defendendo
+                    defendendo = true; // Atribui o valor verdadeiro a variavel defendendo
                     System.out.println("");
                     personagem.defender();
                     break;
@@ -137,10 +237,10 @@ public class Metodos {
                 if (danoRecebido > 0) {
                     personagem.tomarDano(danoRecebido);
                     // Condicional que verifica se o status defendendo é diferente de falso
-                    if (Game.defendendo != false) {
+                    if (defendendo != false) {
                         // Caso seja diferente de falso ele mostra o valor do dano recebido dividido por dois e atribui o valor falso para o status defendendo.
                         System.out.printf("O %s causou %d de dano!\n", monstro.getNome(), (danoRecebido / 2));
-                        Game.defendendo = false; // Atribui o valor falso a variavel defendendo
+                        defendendo = false; // Atribui o valor falso a variavel defendendo
                     } else {
                         System.out.printf("O %s causou %d de dano!\n", monstro.getNome(), danoRecebido);
                     }
