+++ src/classes/Monstro.java
@@ -1,3 +1,4 @@
+package classes;
 import java.util.Random;
 
 public class Monstro {
